+++ src/components/checkbox/index.tsx
@@ -1,31 +1,54 @@
-import React, { PropsWithChildren, useContext, useState } from "react";
+import React, { PropsWithChildren, useContext, useEffect, useState } from "react";
 import { ContextTheme } from "../../provider";
 import { CheckboxStyles } from "./styles";
 import { Pressable, View } from "react-native";
 import { Check } from "lucide-react-native";
 import { Text } from './../text'
+import { IUseForm, WithFormOptional, WithFormRequired } from "../../hook/useForm";
 
-export interface CheckboxProps {
+export interface CheckboxBaseProps {
+  name: string
+  form?: IUseForm
   label?: string
   hint?: string
   defaultIsChecked?: boolean
   onChange?: (isSelected: boolean) => void;
   isDisabled?: boolean
   isInvalid?: boolean;
+  isRequired?: boolean
 }
 
-export const Checkbox = ({ defaultIsChecked = false, label, hint, onChange, isDisabled, isInvalid }: PropsWithChildren<CheckboxProps>): JSX.Element => {
+export type CheckboxProps = CheckboxBaseProps & (WithFormRequired | WithFormOptional);
+
+export const Checkbox = ({ defaultIsChecked = false, isRequired, form, name, label, hint, onChange, isDisabled, isInvalid }: PropsWithChildren<CheckboxProps>): JSX.Element => {
   const [value, setValue] = useState(defaultIsChecked)
   const config = useContext(ContextTheme);
   const styles = CheckboxStyles(config)
 
   const handleChangeValue = () => {
-    onChange && onChange(!value)
+    if (form) {
+      form.handleChange(name, !value)
+    }
+
+    if (onChange) {
+      onChange(!value)
+    }
+
     setValue(!value)
   }
 
+  useEffect(() => {
+    if (form) {
+      form.registerRequired(name, isRequired);
+
+      if (!(name in form.values)) {
+        form.registerField(name, 'boolean');
+      }
+    }
+  }, []);
+
   return (
-    <View>
+    <View style={{ width: '100%' }}>
       <Pressable style={styles.container} onPress={handleChangeValue}>
         <Pressable
           style={[
@@ -38,7 +61,7 @@ export const Checkbox = ({ defaultIsChecked = false, label, hint, onChange, isDi
         >
           {value && (
             <Check
-              color={isInvalid ? config.colors.error : config.colors.secondary}
+              color={(isInvalid || !!form?.errors[name]) ? config.colors.error : config.colors.secondary}
               size={16}
             />
           )}
@@ -48,23 +71,22 @@ export const Checkbox = ({ defaultIsChecked = false, label, hint, onChange, isDi
             style={[
               styles.label,
               isDisabled && styles.labelDisabled,
-              isInvalid && styles.labelInvalid
+              (isInvalid || !!form?.errors[name]) && styles.labelInvalid
             ]}
           >
             {label}
+            {isRequired && <Text style={{ color: config.colors.error }}> *</Text>}
           </Text>
         )}
       </Pressable>
-      {hint && (
-        <Text
-          style={[
-            styles.hint,
-            isInvalid && styles.hintInvalid
-          ]}
-        >
-          {label}
-        </Text>
-      )}
+      <Text
+        style={[
+          styles.hint,
+          (isInvalid || !!form?.errors[name]) && styles.hintInvalid
+        ]}
+      >
+        {form?.errors[name] || hint}
+      </Text>
     </View>
   )
 }