+++ src/components/checkbox/styles.ts
@@ -1,10 +1,11 @@
-import { DefaultTheme } from '../../provider/index';
+import { DefaultTheme, defaultTheme as theme } from '../../provider/index';
 import { StyleSheet } from 'react-native';
 import { ITheme } from '../../provider';
 
 export const CheckboxStyles = (theme: DefaultTheme | ITheme) => StyleSheet.create({
   container: {
-    flexDirection: 'row'
+    flexDirection: 'row',
+    alignSelf: 'flex-start'
   },
   checkbox: {
     alignItems: 'center',