+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New color caution
 - Text Component
 - Heading Component
+- Create hook useForm
 
 ## [1.5.2] - 2024-07-05